Replace Next.js SSR with a static export + hand-rolled request handling
Removes Next.js's server runtime from production entirely (output: 'export'
in next.config.ts) and moves what it used to do at request time — the
proxy.ts auth redirect, environment/brand bootstrap injection, and
next.config.ts's redirects/rewrites/headers — into a small hand-rolled layer
in server.ts/staticFrontend.ts. This eliminates the class of bug that broke
the previous standalone-output build on the Next.js 16.3.1 bump (a file-
tracing gap in @swc/helpers), and decouples the frontend's build graph from
apps/backend's dependencies entirely.

apps/backend is now also its own pnpm workspace package, so the production
Docker image's node_modules can be built via `pnpm deploy` (backend-only
dependencies) instead of the whole monorepo's — the image still grows
slightly net (~1.02GB to ~1.38GB) since Next's own file-tracing was more
precise than this replacement, but recovers most of the ground lost by
dropping standalone tracing altogether (2.49GB naive, 1.82GB hand-trimmed).
