+++ .github/workflows/ci.yml
@@ -23,7 +23,7 @@ jobs:
       - name: Set up pnpm
         uses: pnpm/action-setup@v4
         with:
-          version: 10.10.0
+          version: 10.22.0
 
       - name: Set up Node
         uses: actions/setup-node@v4
@@ -50,7 +50,7 @@ jobs:
       - name: Set up pnpm
         uses: pnpm/action-setup@v4
         with:
-          version: 10.10.0
+          version: 10.22.0
 
       - name: Set up Node
         uses: actions/setup-node@v4
@@ -75,7 +75,7 @@ jobs:
       - name: Set up pnpm
         uses: pnpm/action-setup@v4
         with:
-          version: 10.10.0
+          version: 10.22.0
 
       - name: Set up Node
         uses: actions/setup-node@v4
@@ -142,7 +142,7 @@ jobs:
       - name: Set up pnpm
         uses: pnpm/action-setup@v4
         with:
-          version: 10.10.0
+          version: 10.22.0
 
       - name: Set up Node
         uses: actions/setup-node@v4
