+++ Dockerfile
@@ -25,9 +25,8 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     && ln -sf python3 /usr/bin/python \
     && rm -rf /var/lib/apt/lists/*
 
-RUN npm install -g node-gyp pnpm@10.10.0
+RUN npm install -g node-gyp pnpm@10.22.0
 
-ENV BUILD_STANDALONE=true
 # Temporarily setting the DATABASE_URL to a file in /tmp to ensure accessibility to the db directory during the build process.
 ENV DATABASE_URL=file:///tmp/logicle.sqlite
 # Set pnpm store path in a known position... which we'll mount as a cache volume later
@@ -59,24 +58,46 @@ RUN if [ -n "${APP_VERSION}" ]; then \
       echo 'APP_VERSION not provided; leaving package.json as-is'; \
     fi
 
-# Build the application which also compiles all assets — reuse Next.js cache
+# Build the application which also compiles all assets — reuse Next.js cache.
+# apps/frontend builds to a fully static export (output: 'export' in
+# next.config.ts — no Next server runtime, so nothing left to trace/bundle
+# incorrectly); dist-server (the custom backend + static-file server) is a
+# separate tsup bundle that still needs real node_modules at runtime.
 RUN --mount=type=cache,id=next-cache,target=/app/.next/cache \
     NODE_ENV=production pnpm build
 
-# The custom backend bundle keeps sharp external so its native binary remains
-# loadable. Prepare a compact, dereferenced runtime tree for sharp and its
-# Linux dependencies; Next's standalone output alone does not trace imports
-# made by dist-server.
-RUN sharp_dir="$(readlink -f node_modules/sharp)" \
-    && sharp_deps_dir="$(dirname "$sharp_dir")" \
-    && mkdir -p /app/runtime-node_modules/@img \
-    && cp -aL "$sharp_dir" /app/runtime-node_modules/sharp \
-    && cp -aL "$sharp_deps_dir/@img/colour" /app/runtime-node_modules/@img/colour \
-    && cp -aL "$sharp_deps_dir/@img/sharp-linux-x64" /app/runtime-node_modules/@img/sharp-linux-x64 \
-    && cp -aL "$sharp_deps_dir/@img/sharp-libvips-linux-x64" /app/runtime-node_modules/@img/sharp-libvips-linux-x64 \
-    && cp -aL "$sharp_deps_dir/detect-libc" /app/runtime-node_modules/detect-libc \
-    && cp -aL "$sharp_deps_dir/semver" /app/runtime-node_modules/semver \
-    && tar -C /app/runtime-node_modules -cf /app/runtime-node_modules.tar .
+# apps/backend is its own workspace package (apps/backend/package.json)
+# declaring only what backend code — plus packages/core and
+# packages/file-analyzer, pulled in via tsconfig path aliases and inlined by
+# tsup, not via node_modules — actually imports. `pnpm deploy` resolves a
+# fresh, correctly-pruned node_modules from just that declaration, which
+# excludes the ~50 frontend-only packages (Radix UI, Tailwind, icon packs,
+# CodeMirror, etc.) that a plain `pnpm install --prod` at the repo root
+# would otherwise still include (there being only one shared package.json
+# before this split). `pnpm prune --prod` was tried first instead of a
+# real workspace split and rejected — see git history — it misclassified
+# real runtime deps like better-sqlite3 as devDependencies-safe-to-drop.
+RUN --mount=type=cache,id=pnpm,target=/pnpm/store \
+    pnpm --filter=@logicleai/backend deploy --prod /app/deploy-backend
+
+# `pnpm deploy` resolves apps/backend's own declared dependencies correctly,
+# but a few of those are themselves only reachable via optional/unused
+# integration paths (verified by `grep` over dist-server's actual bundled
+# output, not package.json classification) and safe to drop:
+#  - next + @next/swc-*: only imported dynamically in server.ts, gated on
+#    dev mode (see the `await import('next')` there) — never resolved once
+#    NODE_ENV=production.
+#  - mermaid (+ its own deps mathjax/cytoscape-fcose/mermaid-parser):
+#    exceljs/docx's optional diagram-embedding integration, never actually
+#    invoked by any code path dist-server bundles.
+RUN rm -rf \
+    /app/deploy-backend/node_modules/next \
+    /app/deploy-backend/node_modules/.pnpm/next@* \
+    /app/deploy-backend/node_modules/.pnpm/@next+swc-* \
+    /app/deploy-backend/node_modules/mermaid \
+    /app/deploy-backend/node_modules/.pnpm/mermaid@* \
+    /app/deploy-backend/node_modules/.pnpm/@mermaid-js+* \
+    /app/deploy-backend/node_modules/.pnpm/@mathjax+*
 
 
 # ---------------------
@@ -105,20 +126,16 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     && rm -rf /var/lib/apt/lists/*
 
 # Create and set permissions for directories
-RUN mkdir -p apps/frontend/.next/cache /data/sqlite /data/files \
-    && chown -R node:node apps/frontend/.next /data
+RUN mkdir -p /data/sqlite /data/files \
+    && chown -R node:node /data
 
-# Copy built assets from the 'builder' stage to appropriate locations
-COPY --from=builder /app/apps/frontend/public ./apps/frontend/public
-COPY --from=builder /app/apps/frontend/.next/standalone ./
-COPY --from=builder /app/apps/frontend/.next/static ./apps/frontend/.next/static
+# Copy built assets from the 'builder' stage to appropriate locations.
+# apps/frontend/out is the static export served directly by dist-server
+# (see apps/backend/lib/staticFrontend.ts) — no `.next/standalone` dance.
+COPY --from=builder /app/apps/frontend/out ./apps/frontend/out
 COPY --from=builder /app/dist-server ./dist-server
-COPY --from=builder /app/runtime-node_modules.tar /tmp/runtime-node_modules.tar
-RUN rm -f node_modules/sharp node_modules/detect-libc node_modules/semver \
-    && rm -rf node_modules/@img \
-    && mkdir -p node_modules/@img \
-    && tar -xf /tmp/runtime-node_modules.tar -C node_modules \
-    && rm /tmp/runtime-node_modules.tar
+COPY --from=builder /app/package.json ./package.json
+COPY --from=builder /app/deploy-backend/node_modules ./node_modules
 COPY --from=file-analyzer /mcp-file-analyzer /usr/local/bin/mcp-file-analyzer
 
 # Switch to the non-root 'node' for security reasons
@@ -127,5 +144,6 @@ USER node
 EXPOSE 3000
 
 ENV NODE_ENV=production
-# Start the Next.js standalone server.
+# apps/frontend is a static export served by dist-server itself (see
+# apps/backend/lib/staticFrontend.ts) — there is no separate Next server.
 CMD ["node", "--enable-source-maps", "dist-server/server.js"]
