+++ apps/backend/package.json
@@ -1,0 +1,82 @@
+{
+  "name": "@logicleai/backend",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "dependencies": {
+    "@ai-sdk/anthropic": "3.0.75",
+    "@ai-sdk/google": "^3.0.67",
+    "@ai-sdk/google-vertex": "4.0.121",
+    "@ai-sdk/openai": "3.0.0",
+    "@ai-sdk/perplexity": "3.0.43",
+    "@ai-sdk/provider": "3.0.10",
+    "@ai-sdk/provider-utils": "4.0.26",
+    "@apidevtools/openapi-schemas": "^2.1.0",
+    "@aws-sdk/client-s3": "^3.1044.0",
+    "@aws-sdk/lib-storage": "^3.1044.0",
+    "@google/genai": "^1.52.0",
+    "@google/generative-ai": "^0.24.1",
+    "@libpdf/core": "^0.2.12",
+    "@mjackson/headers": "^0.10.0",
+    "@mjackson/multipart-parser": "^0.7.0",
+    "@modelcontextprotocol/sdk": "^1.29.0",
+    "@node-saml/node-saml": "^5.1.0",
+    "@openmeter/sdk": "1.0.0-beta.227",
+    "@opentelemetry/api": "^1.9.0",
+    "@opentelemetry/api-logs": "^0.203.0",
+    "@opentelemetry/exporter-logs-otlp-http": "^0.203.0",
+    "@opentelemetry/resources": "^2.0.1",
+    "@opentelemetry/sdk-logs": "^0.203.0",
+    "@opentelemetry/sdk-trace-base": "^2.1.0",
+    "@opentelemetry/semantic-conventions": "^1.37.0",
+    "@opentelemetry/winston-transport": "^0.14.1",
+    "@readme/openapi-parser": "^2.6.0",
+    "@smithy/node-http-handler": "^4.7.5",
+    "@vercel/otel": "^2.0.1",
+    "ai": "6.0.175",
+    "ajv": "^8.17.1",
+    "ajv-draft-04": "^1.0.0",
+    "bcryptjs": "3.0.3",
+    "better-sqlite3": "^12.6.2",
+    "docx": "9.5.1",
+    "exceljs": "^4.4.0",
+    "fast-xml-parser": "^5.7.3",
+    "form-data": "^4.0.5",
+    "google-auth-library": "^10.6.2",
+    "iron-session": "^8.0.4",
+    "js-tiktoken": "^1.0.8",
+    "json-schema": "^0.4.0",
+    "kysely": "^0.28.17",
+    "lru-cache": "^11.0.2",
+    "mammoth": "^1.10.0",
+    "meilisearch": "^0.41.0",
+    "micromatch": "^4.0.5",
+    "mime-types": "^3.0.1",
+    "nanoid": "^5.0.4",
+    "next": "^16.3.1",
+    "openai": "^6.15.0",
+    "openapi-types": "^12.1.3",
+    "openid-client": "^6.8.1",
+    "openpgp": "^6.1.1",
+    "pg": "^8.14.1",
+    "pptx2json": "^0.0.10",
+    "remark": "^15.0.1",
+    "remark-docx": "^0.3.26",
+    "remark-gfm": "4.0",
+    "remark-math": "^6.0.0",
+    "remark-parse": "^11.0.0",
+    "replicate": "^1.4.0",
+    "sharp": "^0.34.5",
+    "together-ai": "^0.39.0",
+    "turndown": "^7.2.1",
+    "turndown-plugin-gfm": "^1.0.2",
+    "undici": "^7.25.0",
+    "unified": "^11.0.5",
+    "unist-util-visit": "^5.0.0",
+    "winston": "^3.15.0",
+    "ws": "^8.18.3",
+    "xlsx": "file:../../vendor/xlsx",
+    "yaml": "^2.6.0",
+    "zod": "^4.0.0"
+  }
+}
