+++ apps/frontend/app/admin/backends/[id]/PageClient.tsx
@@ -1,0 +1,39 @@
+'use client'
+import { useBackend } from '@/hooks/backends'
+import { useRouter } from 'next/navigation'
+import BackendForm, { BackendFormFields } from '../components/BackendForm'
+import { mutate } from 'swr'
+import toast from 'react-hot-toast'
+import { useTranslation } from 'react-i18next'
+import { patch } from '@/lib/fetch'
+import { AdminPage } from '../../components/AdminPage'
+import { useUrlSegment } from '@/hooks/useUrlSegment'
+
+const BackendPage = () => {
+  const id = useUrlSegment(2)
+  const { t } = useTranslation()
+  const { isLoading, error, data: backend } = useBackend(id)
+  const router = useRouter()
+
+  async function onSubmit(values: Partial<BackendFormFields>) {
+    const url = `/api/backends/${id}`
+    const response = await patch(url, { providerType: backend?.providerType, ...values })
+    if (response.error) {
+      toast.error(response.error.message)
+      return
+    }
+    await mutate(url)
+    toast.success(t('backend-successfully-updated'))
+    router.push(`/admin/backends`)
+  }
+
+  return (
+    <AdminPage isLoading={isLoading} error={error} title={`Backend ${backend?.name}`}>
+      {backend && (
+        <BackendForm backend={backend as unknown as BackendFormFields} onSubmit={onSubmit} />
+      )}
+    </AdminPage>
+  )
+}
+
+export default BackendPage
