+++ apps/frontend/app/admin/backends/[id]/page.tsx
@@ -1,38 +1,11 @@
-'use client'
-import { useBackend } from '@/hooks/backends'
-import { useParams, useRouter } from 'next/navigation'
-import BackendForm, { BackendFormFields } from '../components/BackendForm'
-import { mutate } from 'swr'
-import toast from 'react-hot-toast'
-import { useTranslation } from 'react-i18next'
-import { patch } from '@/lib/fetch'
-import { AdminPage } from '../../components/AdminPage'
+// Thin server wrapper: output:'export' requires generateStaticParams on any
+// file in the dynamic-segment path, and that export can't live in a 'use
+// client' module. The actual id is read client-side (useParams(), against
+// the live URL) in PageClient — this placeholder value is never used.
+import PageClient from './PageClient'
 
-const BackendPage = () => {
-  const { id } = useParams() as { id: string }
-  const { t } = useTranslation()
-  const { isLoading, error, data: backend } = useBackend(id)
-  const router = useRouter()
+export const generateStaticParams = () => [{ id: '_' }]
 
-  async function onSubmit(values: Partial<BackendFormFields>) {
-    const url = `/api/backends/${id}`
-    const response = await patch(url, { providerType: backend?.providerType, ...values })
-    if (response.error) {
-      toast.error(response.error.message)
-      return
-    }
-    await mutate(url)
-    toast.success(t('backend-successfully-updated'))
-    router.push(`/admin/backends`)
-  }
-
-  return (
-    <AdminPage isLoading={isLoading} error={error} title={`Backend ${backend?.name}`}>
-      {backend && (
-        <BackendForm backend={backend as unknown as BackendFormFields} onSubmit={onSubmit} />
-      )}
-    </AdminPage>
-  )
+export default function Page() {
+  return <PageClient />
 }
-
-export default BackendPage
