+++ apps/frontend/app/admin/users/[userId]/page.tsx
@@ -1,12 +1,11 @@
-'use client'
+// Thin server wrapper: output:'export' requires generateStaticParams on any
+// file in the dynamic-segment path, and that export can't live in a 'use
+// client' module. The actual id is read client-side (useParams(), against
+// the live URL) in PageClient — this placeholder value is never used.
+import PageClient from './PageClient'
 
-import { useParams } from 'next/navigation'
-import { User } from '../components/User'
+export const generateStaticParams = () => [{ userId: '_' }]
 
-const UserPage = () => {
-  const params = useParams() as { userId: string }
-  const { userId } = params
-  return <User userId={userId}></User>
+export default function Page() {
+  return <PageClient />
 }
-
-export default UserPage
