+++ apps/frontend/app/context/environmentProvider.tsx
@@ -3,6 +3,7 @@ import { LlmModel } from '@/lib/chat/models'
 import { useContext } from 'react'
 import React from 'react'
 import * as dto from '@/types/dto'
+import { ENVIRONMENT_ELEMENT_ID, readBootstrapJson } from '@/lib/bootstrapPlaceholders'
 
 export type Environment = {
   appUrl: string
@@ -35,11 +36,16 @@ export type Environment = {
 export const EnvironmentContext = React.createContext<Environment>({} as Environment)
 
 type Props = {
-  value: Environment
   children: React.ReactNode
 }
 
-export const EnvironmentProvider: React.FC<Props> = ({ children, value }) => {
+// server.ts splices the real payload into a <script id="__logicle_env__">
+// tag before sending the HTML, so it's already in the DOM by the time this
+// component's lazy initializer runs on the client — no fetch, no waterfall.
+export const EnvironmentProvider: React.FC<Props> = ({ children }) => {
+  const [value] = React.useState<Environment>(
+    () => readBootstrapJson<Environment>(ENVIRONMENT_ELEMENT_ID) ?? ({} as Environment)
+  )
   return <EnvironmentContext.Provider value={value}>{children}</EnvironmentContext.Provider>
 }
